big update, functionalities and frontend done

+++ src/main/java/eu/lukks/domain/ReservationDto.java
@@ -17,8 +17,8 @@ import lombok.Setter;
 @NoArgsConstructor
 public class ReservationDto {
 
-private Long id;
-	
+	private Long id;
+
 	private String name;
 	private String surname;
 	private String address;
@@ -26,16 +26,16 @@ private Long id;
 	private String city;
 	private String phone;
 	private String mail;
-	private Integer price;
+	private Double price;
 	private String roomTitle;
 	private Long roomId;
 	private Boolean breakfast;
 	private Boolean parking;
-	
+
 	@DateTimeFormat(pattern = "yyyy-MM-dd")
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
 	private LocalDate dateFrom;
-	
+
 	@DateTimeFormat(pattern = "yyyy-MM-dd")
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
 	private LocalDate dateTo;
